+++ src/entities/schemas/webhook_data/nested_schemas.py
@@ -127,70 +127,6 @@ class FromTo(BaseModel):
     to: str | None
 
 
-class DiffMilestone(FromTo):
-    pass
-
-
-class DiffSprintOrder(BaseModel):
-    pass
-
-
-class DiffDueDate(BaseModel):
-    pass
-
-
-class DiffStatus(FromTo):
-    pass
-
-
-class DiffBlockedNote(FromTo):
-    pass
-
-
-class DiffBlockedNoteHTML(FromTo):
-    pass
-
-
-class IsBlocked(FromTo):
-    pass
-
-
-class IsIocaine(FromTo):
-    pass
-
-
-class AssignedTo(FromTo):
-    pass
-
-
-class Name(FromTo):
-    pass
-
-
-class EstimatedStart(FromTo):
-    pass
-
-
-class EstimatedFinish(FromTo):
-    pass
-
-
-class TeamRequirements(FromTo):
-    pass
-
-
-class ClientRequirements(FromTo):
-    pass
-
-
-class DiffTypePrioritySeverity(FromTo):
-    pass
-
-
-class Points(BaseModel):
-    __root__: dict[str, FromTo]
-
-
 class UserStory(BaseItem):
     """
     Nested from BaseItem:
